requirements.txt
conan>=1.42.0
joulescope_ui>=0.9.0
intelhex>=2.3.0
